requirements.txt
anyio==3.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
instagrapi==1.18.1
Pillow==10.0.0
pycryptodomex==3.18.0
pydantic==1.10.9
PySocks==1.7.1
python-telegram-bot==20.3
requests==2.31.0
requests-toolbelt==1.0.0
sniffio==1.3.0
socksio==1.0.0
typing_extensions==4.7.1
urllib3==2.0.3
